odev/requirements.txt
appdirs==1.4.4
asgiref==3.2.10
distlib==0.3.1
Django==3.0.8
django-admin-honeypot==1.1.0
django-admin-sortable==2.2.3
django-ckeditor==5.9.0
django-cors-headers==3.4.0
django-environ==0.4.5
django-js-asset==1.2.2
django-modeltranslation==0.15.1
django-mssql-backend==2.8.1
djangorestframework==3.11.0
filelock==3.0.12
Pillow==7.2.0
pyodbc==4.0.30
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
virtualenv==20.0.29
wfastcgi==3.0.0
